Extract brush trace query helpers
